Hack to fix GC mutex after a fork.
Even though after a fork the child only inherits the single thread that called
the fork(), if another thread in the parent was attempting to lock the mutex
while being held in fork_child_prepare(), the mutex will be left in a
inconsistent state in the child after the UNLOCK(). This is the case at least in
Mac OS X and leads to an unusable GC in the child which will block when
attempting to perform any GC operation that acquires the mutex.